[VPN 2.0] Move credential management utils to V1 service helper
In VPN architecture 2.0, we are planning to implement our own
credential management code. Hence, these functions should be
moved from common utilities to v1 browser code. A couple of
those functions is used in v1 subcomponents (connection and
wireguard), so they have been extracted to a separate target.

Implements part of https://github.com/brave/brave-browser/issues/56692
