+++ components/brave_vpn/browser/brave_vpn_service_helper.cc
@@ -107,4 +107,80 @@ base::Time GetExpirationTimeForSkusCredential(PrefService* local_prefs) {
   return *base::ValueToTime(expiration_time_value);
 }
 
+bool HasValidSubscriberCredential(PrefService* local_prefs) {
+  const base::DictValue& sub_cred_dict =
+      local_prefs->GetDict(prefs::kBraveVPNSubscriberCredential);
+  if (sub_cred_dict.empty()) {
+    return false;
+  }
+
+  const std::string* cred = sub_cred_dict.FindString(kSubscriberCredentialKey);
+  const base::Value* expiration_time_value =
+      sub_cred_dict.Find(kSubscriberCredentialExpirationKey);
+
+  if (!cred || !expiration_time_value) {
+    return false;
+  }
+
+  if (cred->empty()) {
+    return false;
+  }
+
+  auto expiration_time = base::ValueToTime(expiration_time_value);
+  if (!expiration_time || expiration_time < base::Time::Now()) {
+    return false;
+  }
+
+  return true;
+}
+
+std::string GetSubscriberCredential(PrefService* local_prefs) {
+  if (!HasValidSubscriberCredential(local_prefs)) {
+    return "";
+  }
+  const base::DictValue& sub_cred_dict =
+      local_prefs->GetDict(prefs::kBraveVPNSubscriberCredential);
+  const std::string* cred = sub_cred_dict.FindString(kSubscriberCredentialKey);
+  DCHECK(cred);
+  return *cred;
+}
+
+bool HasValidSkusCredential(PrefService* local_prefs) {
+  const base::DictValue& sub_cred_dict =
+      local_prefs->GetDict(prefs::kBraveVPNSubscriberCredential);
+  if (sub_cred_dict.empty()) {
+    return false;
+  }
+
+  const std::string* skus_cred = sub_cred_dict.FindString(kSkusCredentialKey);
+  const base::Value* expiration_time_value =
+      sub_cred_dict.Find(kSubscriberCredentialExpirationKey);
+
+  if (!skus_cred || !expiration_time_value) {
+    return false;
+  }
+
+  if (skus_cred->empty()) {
+    return false;
+  }
+
+  auto expiration_time = base::ValueToTime(expiration_time_value);
+  if (!expiration_time || expiration_time < base::Time::Now()) {
+    return false;
+  }
+
+  return true;
+}
+
+std::string GetSkusCredential(PrefService* local_prefs) {
+  CHECK(HasValidSkusCredential(local_prefs))
+      << "Don't call when there is no valid skus credential.";
+
+  const base::DictValue& sub_cred_dict =
+      local_prefs->GetDict(prefs::kBraveVPNSubscriberCredential);
+  const std::string* skus_cred = sub_cred_dict.FindString(kSkusCredentialKey);
+  DCHECK(skus_cred);
+  return *skus_cred;
+}
+
 }  // namespace brave_vpn
