+++ components/brave_vpn/browser/connection/ikev2/BUILD.gn
@@ -15,6 +15,7 @@ source_set("ikev2") {
   ]
   deps = [
     "//base",
+    "//brave/components/brave_vpn/browser:v1_helper",
     "//brave/components/brave_vpn/browser/api",
     "//brave/components/brave_vpn/browser/connection:api",
     "//brave/components/brave_vpn/browser/connection:common",
