+++ components/brave_vpn/browser/connection/wireguard/wireguard_connection_api_impl_base.cc
@@ -8,6 +8,7 @@
 #include "base/check.h"
 #include "base/logging.h"
 #include "brave/components/brave_vpn/browser/api/brave_vpn_api_request.h"
+#include "brave/components/brave_vpn/browser/brave_vpn_service_helper.h"
 #include "brave/components/brave_vpn/browser/connection/brave_vpn_connection_manager.h"
 #include "brave/components/brave_vpn/browser/connection/brave_vpn_region_data_manager.h"
 #include "brave/components/brave_vpn/common/brave_vpn_utils.h"
