+++ components/brave_vpn/common/brave_vpn_utils.cc
@@ -10,8 +10,6 @@
 #include "base/check.h"
 #include "base/containers/fixed_flat_map.h"
 #include "base/feature_list.h"
-#include "base/json/json_writer.h"
-#include "base/json/values_util.h"
 #include "base/notreached.h"
 #include "base/strings/string_split.h"
 #include "brave/components/brave_vpn/common/brave_vpn_constants.h"
@@ -319,82 +317,6 @@ void MigrateLocalStatePrefs(PrefService* local_prefs) {
 
 void ClearSubscriberCredential(PrefService* local_prefs) {
   local_prefs->ClearPref(prefs::kBraveVPNSubscriberCredential);
-}
-
-bool HasValidSubscriberCredential(PrefService* local_prefs) {
-  const base::DictValue& sub_cred_dict =
-      local_prefs->GetDict(prefs::kBraveVPNSubscriberCredential);
-  if (sub_cred_dict.empty()) {
-    return false;
-  }
-
-  const std::string* cred = sub_cred_dict.FindString(kSubscriberCredentialKey);
-  const base::Value* expiration_time_value =
-      sub_cred_dict.Find(kSubscriberCredentialExpirationKey);
-
-  if (!cred || !expiration_time_value) {
-    return false;
-  }
-
-  if (cred->empty()) {
-    return false;
-  }
-
-  auto expiration_time = base::ValueToTime(expiration_time_value);
-  if (!expiration_time || expiration_time < base::Time::Now()) {
-    return false;
-  }
-
-  return true;
-}
-
-std::string GetSubscriberCredential(PrefService* local_prefs) {
-  if (!HasValidSubscriberCredential(local_prefs)) {
-    return "";
-  }
-  const base::DictValue& sub_cred_dict =
-      local_prefs->GetDict(prefs::kBraveVPNSubscriberCredential);
-  const std::string* cred = sub_cred_dict.FindString(kSubscriberCredentialKey);
-  DCHECK(cred);
-  return *cred;
-}
-
-bool HasValidSkusCredential(PrefService* local_prefs) {
-  const base::DictValue& sub_cred_dict =
-      local_prefs->GetDict(prefs::kBraveVPNSubscriberCredential);
-  if (sub_cred_dict.empty()) {
-    return false;
-  }
-
-  const std::string* skus_cred = sub_cred_dict.FindString(kSkusCredentialKey);
-  const base::Value* expiration_time_value =
-      sub_cred_dict.Find(kSubscriberCredentialExpirationKey);
-
-  if (!skus_cred || !expiration_time_value) {
-    return false;
-  }
-
-  if (skus_cred->empty()) {
-    return false;
-  }
-
-  auto expiration_time = base::ValueToTime(expiration_time_value);
-  if (!expiration_time || expiration_time < base::Time::Now()) {
-    return false;
-  }
-
-  return true;
-}
-
-std::string GetSkusCredential(PrefService* local_prefs) {
-  CHECK(HasValidSkusCredential(local_prefs))
-      << "Don't call when there is no valid skus credential.";
-
-  const base::DictValue& sub_cred_dict =
-      local_prefs->GetDict(prefs::kBraveVPNSubscriberCredential);
-  const std::string* skus_cred = sub_cred_dict.FindString(kSkusCredentialKey);
-  DCHECK(skus_cred);
-  return *skus_cred;
 }
 
 }  // namespace brave_vpn
